+++ next/api/src/router/customer-service.ts
@@ -4,8 +4,8 @@ import { auth, customerServiceOnly } from '../middleware/auth';
 import { Group } from '../model/Group';
 import { Role } from '../model/Role';
 import { User } from '../model/User';
-import { GroupResponse } from '../json/group';
-import { UserResponse } from '../json/user';
+import { GroupResponse } from '../response/group';
+import { UserResponse } from '../response/user';
 
 const router = new Router().use(auth, customerServiceOnly);
